refactoring entities and names
