+++ dashboard-notifications/src/main/java/uk.gov.hmcts.reform/dashboard/repositories/NotificationTemplateRepository.java
@@ -1,8 +1,8 @@
-package uk.gov.hmcts.reform.dashboard.repository;
+package uk.gov.hmcts.reform.dashboard.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import uk.gov.hmcts.reform.dashboard.data.NotificationTemplateEntity;
+import uk.gov.hmcts.reform.dashboard.entities.NotificationTemplateEntity;
 import javax.transaction.Transactional;
 
 @Transactional
