+++ dashboard-notifications/src/main/java/uk.gov.hmcts.reform/dashboard/services/TaskListService.java
@@ -1,10 +1,10 @@
-package uk.gov.hmcts.reform.dashboard.service;
+package uk.gov.hmcts.reform.dashboard.services;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.dashboard.data.TaskListEntity;
-import uk.gov.hmcts.reform.dashboard.repository.TaskListRepository;
+import uk.gov.hmcts.reform.dashboard.entities.TaskListEntity;
+import uk.gov.hmcts.reform.dashboard.repositories.TaskListRepository;
 
 import java.util.List;
 import java.util.Optional;
